requirements.txt
streamlit==1.45.0
groq==0.28.0
python-dotenv==1.1.0